changed header size to 16 bit

+++ src/main.cpp
@@ -3,15 +3,15 @@
 #include <fstream>
 
 
-std::vector<i16> loadFile(std::string fileName) {
+std::vector<u16> loadFile(std::string fileName) {
     std::ifstream file(fileName, std::ios::in | std::ios::binary);
-    size_t size;
-    std::vector<i16> program;
+    u16 size;
+    std::vector<u16> program;
     file.read((char*) &size, sizeof(size));
     program.reserve(size);
     program.resize(size);
     file.seekg(sizeof(size_t));
-    file.read((char*) &program[0], size * sizeof(i16));
+    file.read((char*) &program[0], size * sizeof(u16));
     file.close();
     
     return program;
@@ -23,7 +23,7 @@ int main(int argc, char** argv) {
         return 1;
     }
 
-    std::vector<i16> program = loadFile(argv[1]);
+    std::vector<u16> program = loadFile(argv[1]);
     VM vm;
     vm.loadProgram(program);
     vm.run();
